+++ documentation/components/Logo.tsx
@@ -4,10 +4,12 @@ export type LogoProps = {
   hoverEffect?: boolean;
 };
 
+import LogoSvg from "../public/secsync-logo.svg";
+
 export const Logo = ({ color = "currentColor", height = 20 }: LogoProps) => {
   return (
     <div className="relative" style={{ height: height }}>
-      Secsync
+      <LogoSvg width={160} />
     </div>
   );
 };